+++ docs/whatsnew.md
@@ -11,6 +11,7 @@
 - new members 'toIEC' and 'fromIEC' for Variable object
 - new property 'cpuInfoExtended' for Cpu object
 - new properties 'moduleInfo' and 'moduleInfoExtended' for Module objects
+- new member 'delete' for Module objects
 - new member function 'readRequestStatus' for all objects
 
 ## V1.2.9